fix(benv): add home link to /install (#382)
fix breakpad, which seems to use the include path in the directory it was installed into.

The build broke when we moved the compiled libraries to /install in the benv. Fix by adding a link in $HOME/install to /install.

Verified by compiling a benv and running CI builds with that container as vars.BENV_IMAGE