build: add cmake defaults so users can build without build.sh (#398)
We have build.sh in the build environment that supports multiple configuration switches for the build (debug, asan, etc.), but this can be cumbersome for developers.

Add cmake defaults so users can enter in a build directory and run cmake <repo path> to configure a debug build with reasonable defaults.

The build still assumes it is running inside the build environment (that library dependencies have been installed in /install)