feat(build): refactor version info to cc file to speed up build (#399)
cmake materializes the version into an h file, used in different printouts throughout the code. This h file is widely included in the build, yielding ccache misses and slowing down incremental builds. The version included a patch number derived from the commit tree (so it changes on every commit).

Refactor version info into a library versions, only it will have to recompile as the version changes, and libraries depending on it linked, which would speed up ccache/incremental builds.