feat(build): add build environment multi-arch builds (#400)
Add arm64 build environment container. This should help publish arm64 images in the future, and serve as a devcontainer on macos.

Refactor the current build environment's github workflow into a reusable workflow build-benv-single-arch.yaml that receives the runner and arch as inputs.

Modify the build-tools-single-stage/action.yml to accept the arch and encode it in the image name.

Add build-benv-multiarch.yaml that calls the reusable workflow and publishes a merged manifest to container repos.