feat(render): Add render_parser library
Add Parser struct that a user can configure with MessageMetadata. Then,
given a serialized buffer, the Parser detects the message and its length,
and returns a slice with the message's bytes and a value that the user
configured when it added the message. Users can use that value to store
a handler function for the message.

This will be used to parse render messages in Rust, in lieu of render's
generated protocol.{h,cc} and transform_builder.{h,cc}

(cherry picked from commit 3bc1b65da40487c65b9b8e867c8b56b91ce26ce4)